update user role and permission

+++ app/api/collections/[collectionId]/route.ts
@@ -1,3 +1,4 @@
+import Admin from "@/lib/models/Admin";
 import Collection from "@/lib/models/Collection";
 import Product from "@/lib/models/Product";
 import { connectToDB } from "@/lib/mongoDB";
@@ -44,6 +45,13 @@ export const POST = async (
 
     await connectToDB();
 
+    const admin = await Admin.find({ clerkId: userId });
+    if (admin[0]?.role !== "super_admin") {
+      return new NextResponse("You are not allowed to do this operation", {
+        status: 401,
+      });
+    }
+
     let collection = await Collection.findById(params.collectionId);
 
     if (!collection) {
@@ -84,6 +92,13 @@ export const DELETE = async (
     }
 
     await connectToDB();
+
+    const admin = await Admin.find({ clerkId: userId });
+    if (admin[0]?.role !== "super_admin") {
+      return new NextResponse("You are not allowed to do this operation", {
+        status: 401,
+      });
+    }
 
     await Collection.findByIdAndDelete(params.collectionId);
